requirements.txt

python-docx
pandas
xlsxwriter
streamlit==1.25.0      # Streamlit for building the app
requests==2.31.0       # HTTP requests library
PyPDF2==3.0.1          # PDF parsing
langdetect==1.0.9      # Language detection
tenacity==8.2.3        # Retry logic
python-dotenv==1.0.0   # Environment variable handling
